cos/queue_ops: extract flow accounting + drain orchestration (#1034 P2)
Two more sibling submodules under cos/queue_ops/:

**accounting.rs** (95 LOC) — per-flow byte counter helpers called from
push/pop hot paths:
- account_cos_queue_flow_enqueue
- account_cos_queue_flow_dequeue

**drain.rs** (99 LOC) — bulk-empty + rollback orchestration:
- cos_queue_drain_all (worker exit / RG demote teardown)
- cos_queue_restore_front (partial-commit rollback path)
- cos_queue_clear_orphan_snapshot_after_drop (pop-snapshot stack
  cleanup when a previously-popped item is dropped not retried)

queue_ops/mod.rs production LOC (excluding mod tests):
  4,370 → 4,203 (Phase 2 of 5)

Visibility:
- accounting fns kept private (only called from push/pop in mod.rs);
  imported via `use accounting::{...};`
- drain fns re-exported at pub(in crate::afxdp) — they're the ones
  cos/mod.rs already lists in the public re-export block

build clean, 865 Rust tests pass. Cluster smoke required (touches
the worker drain hot path).
